+++ CHANGES.md
@@ -2,16 +2,15 @@
 
 ## 2.92.0
 
-This release supports the new [`--uploaded-prior-to`](https://pip.pypa.io/en/stable/user_guide/#filtering-by-upload-time).
+This release supports Pip's new [`--uploaded-prior-to`](https://pip.pypa.io/en/stable/user_guide/#filtering-by-upload-time).
 
-* Pass thru `--uploaded-prior-to` on Pip v26.0. (#3131)
+* Support `--uploaded-prior-to` for Pip v26.0 and newer. (#3131)
 
 ## 2.91.7
 
 This release plugs holes in the WHEEL metadata fix in 2.91.6.
 
 * Thread .whl name / platform tag to Wheel parse. (#3137)
->>>>>>> origin/main
 
 ## 2.91.6